Better tests and visualization on 001

+++ 001/001.cpp
@@ -3,23 +3,21 @@
 #define LIMIT 1000
 
 int main() {
-    // This number will never be passed with this algorithm
-    unsigned result = 5;
-    for (unsigned i = 3; i < LIMIT; i += 6) {
-        unsigned j = i;
+    unsigned long long result = 5;
+    for (unsigned long long i = 3; i < LIMIT; i += 6) {
+        unsigned long long j = i;
         while (j < LIMIT) {
             result += j;
             if (!(j % 5)) {
                 result += (j-5);
-                if (j+5 < LIMIT) {
+                if (j + 5 < LIMIT) {
                     result += (j+5);
                 }
             }
-            //printf("result = %u, i = %u\n", result, j);
             j = j << 1;
         }
     }
 
-    printf("Result is %u\n", result);
+    printf("Result is %llu\n", result);
     return 0;
 }
